fix: support Capacitor 1 apps
The Ionic CLI attempts to get the users Capacitor config with the `npx cap config --json` command that was introduced in Capacitor 3. This code assumes that if the command does not work then an output (stdout) would not be returned. This is the case with Capacitor 2 as it returns the "command not found" error to stderror. However, Capacitor 1 returns this to stdout which then causes an error because there is an output, but it not JSON and not parsable.